requirements.txt
Pillow>=10.0.0
numpy>=1.24.0
pyautogui>=0.9.54
